TCPコネクション越しに動くようにした。 ちょっと実装が汚い気もする。
